break apart BrowserRuntime class to be more clear
