+++ packages/common/src/browser/extension.ts
@@ -1,0 +1,79 @@
+import { BrowserRuntimeCommon } from "./common";
+
+//
+// Browser apis that can be used on extension only.
+//
+export class BrowserRuntimeExtension {
+  public static getUrl(scriptName: string): string {
+    return chrome
+      ? chrome.runtime.getURL(scriptName)
+      : browser.runtime.getURL(scriptName);
+  }
+
+  public static sendMessageActiveTab(msg: any) {
+    return chrome
+      ? chrome.tabs.query({ active: true, currentWindow: true }, ([tab]) => {
+          if (tab?.id) chrome.tabs.sendMessage(tab.id, msg);
+        })
+      : browser.tabs
+          .query({ active: true, currentWindow: true })
+          .then(([tab]) => {
+            if (tab?.id) browser.tabs.sendMessage(tab.id, msg);
+          });
+  }
+
+  public static sendMessageTab(tabId: number, msg: any) {
+    chrome
+      ? chrome.tabs.sendMessage(tabId, msg)
+      : browser.tabs.sendMessage(tabId, msg);
+  }
+
+  public static async openWindow(options: chrome.windows.CreateData) {
+    return new Promise((resolve, reject) => {
+      // TODO: `browser` support
+      chrome?.windows.create(options, (newWindow) => {
+        const error = BrowserRuntimeCommon.checkForError();
+        if (error) {
+          return reject(error);
+        }
+        return resolve(newWindow);
+      });
+    });
+  }
+
+  public static async getLastFocusedWindow(): Promise<chrome.windows.Window>;
+  public static async getLastFocusedWindow(): Promise<browser.windows.Window>;
+  public static async getLastFocusedWindow() {
+    return new Promise((resolve) => {
+      chrome
+        ? chrome.windows.getLastFocused(resolve)
+        : browser.windows.getLastFocused().then(resolve);
+    });
+  }
+
+  public static activeTab(): Promise<chrome.tabs.Tab>;
+  public static activeTab(): Promise<browser.tabs.Tab>;
+  public static activeTab() {
+    return new Promise((resolve) => {
+      chrome
+        ? chrome.tabs.query({ active: true, currentWindow: true }, ([tab]) => {
+            resolve(tab);
+          })
+        : browser.tabs
+            .query({ active: true, currentWindow: true })
+            .then(([tab]) => {
+              resolve(tab);
+            });
+    });
+  }
+
+  public static closeActiveTab(): void {
+    chrome
+      ? chrome.tabs.getCurrent((tab) => {
+          if (tab?.id) chrome.tabs.remove(tab.id, function () {});
+        })
+      : browser.tabs.getCurrent().then((tab) => {
+          if (tab?.id) browser.tabs.remove(tab.id);
+        });
+  }
+}
