+++ packages/common/src/browser/index.ts
@@ -1,0 +1,2 @@
+export * from "./common";
+export * from "./extension";
