+++ packages/extension/src/contentScript/index.ts
@@ -1,7 +1,7 @@
 import {
   getLogger,
   ChannelContentScript,
-  BrowserRuntime,
+  BrowserRuntimeExtension,
   CHANNEL_RPC_REQUEST,
   CHANNEL_RPC_RESPONSE,
   CHANNEL_NOTIFICATION,
@@ -27,7 +27,7 @@ function injectScript(scriptName: string) {
     const container = document.head || document.documentElement;
     const scriptTag = document.createElement("script");
     scriptTag.setAttribute("async", "false");
-    scriptTag.src = BrowserRuntime.getUrl(scriptName);
+    scriptTag.src = BrowserRuntimeExtension.getUrl(scriptName);
     container.insertBefore(scriptTag, container.children[0]);
     container.removeChild(scriptTag);
   } catch (error) {
